# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PinNode HTTP API
  version: 1.0.0
  description: Current clean v1 contract implemented by the PinNode server.
  license:
    name: BSD 3-Clause
    identifier: BSD-3-Clause
servers:
  - url: https://pinnode.example
security: []
...
paths:
...
  /v1/sessions:
    get:
      summary: List session history
      tags: [Sessions]
      operationId: listSessions
      security:
        - adminCookie: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Session history, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/HistoricalSession"
        "4XX":
          $ref: "#/components/responses/Error"
        "5XX":
          $ref: "#/components/responses/Error"
    post:
      summary: Create a session idempotently
      tags: [Sessions]
      operationId: createSession
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSessionRequest"
      responses:
        "201":
          description: New session, or an identical encrypted replay for the same key and request.
          headers:
            Idempotent-Replayed:
              description: Present and true only for a replay.
              schema: {type: string, const: "true"}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StartSessionResponse"
        "4XX":
          $ref: "#/components/responses/Error"
        "5XX":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    adminCookie:
      type: apiKey
      in: cookie
      name: pinnode_admin_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: PinNode session token
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: "^[!-~]+$"
...
  responses:
    Error:
      description: Structured API error.
      headers:
        X-Request-ID:
          required: true
          schema: {type: string}
        Retry-After:
          description: Seconds until retry; present on rate limits.
          schema: {type: integer}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    NullableTime:
      type: [string, "null"]
      format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, requestId]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable]
          properties:
            code: {type: string, minLength: 1}
            message: {type: string}
            retryable: {type: boolean}
        requestId: {type: string, minLength: 1}
...
    ExitPolicyPatch:
      type: object
      additionalProperties: false
      properties:
        onAppClose: {type: boolean}
        networkChange: {type: string, enum: ["", any-change, wifi-lost, cellular-lost]}
        afterConfigSeconds: {type: integer, minimum: 0, maximum: 31536000}
        afterLoginSeconds: {type: integer, minimum: 0, maximum: 31536000}
        at: {type: string}
    ExitPolicy:
      allOf:
        - $ref: "#/components/schemas/ExitPolicyPatch"
        - type: object
          required: [onAppClose, networkChange, afterConfigSeconds, afterLoginSeconds, at]
    SessionConfigPatch:
      type: object
      additionalProperties: false
      properties:
        networkMode: {type: string, enum: ["default", "cellular"]}
        vpnEnabled: {type: boolean}
        acceptRoutes: {type: boolean}
        acceptDNS: {type: boolean}
        tailscaleIp: {type: string, maxLength: 15, description: Optional Tailscale IPv4 assigned after the device joins.}
        useExitNode: {type: boolean}
        exitNodeId: {type: string}
        exitNodeIp: {type: string}
        autoExitNode: {type: string, enum: ["", "auto:any"]}
        exitNodeAllowLanAccess: {type: boolean}
        subnetRouter: {type: boolean}
        autoGatewayRoute: {type: boolean}
        autoWiFiSubnetRoute: {type: boolean}
        advertiseRoutes:
          type: array
          maxItems: 16
          items: {type: string}
        advertiseExitNode: {type: boolean}
        disableSNAT: {type: boolean}
        noStatefulFiltering: {type: boolean}
        shieldsUp: {type: boolean}
        runSSHServer: {type: boolean}
        runWebClient: {type: boolean}
        postureChecking: {type: boolean}
        remoteConfig: {type: boolean}
        hostname: {type: string, maxLength: 255}
        netfilterMode: {type: string, enum: ["", "on", "off", "nodivert"]}
        appConnector: {type: boolean}
        exitPolicy: {$ref: "#/components/schemas/ExitPolicyPatch"}
    SessionConfig:
      allOf:
        - $ref: "#/components/schemas/SessionConfigPatch"
        - type: object
          required:
            - networkMode
            - vpnEnabled
            - acceptRoutes
            - acceptDNS
            - tailscaleIp
            - useExitNode
            - exitNodeId
            - exitNodeIp
            - autoExitNode
            - exitNodeAllowLanAccess
            - subnetRouter
            - autoGatewayRoute
            - autoWiFiSubnetRoute
            - advertiseRoutes
            - advertiseExitNode
            - disableSNAT
            - noStatefulFiltering
            - shieldsUp
            - runSSHServer
            - runWebClient
            - postureChecking
            - remoteConfig
            - hostname
            - netfilterMode
            - appConnector
            - exitPolicy
          properties:
            exitPolicy: {$ref: "#/components/schemas/ExitPolicy"}
...
    StartSessionRequest:
      type: object
      additionalProperties: false
      required: [code]
      properties:
        code: {type: string, pattern: "^[0-9]{6}$"}
        gatewayRoute: {type: string, default: ""}
        wifiSubnetRoute: {type: string, default: ""}
    StartSessionResponse:
      type: object
      additionalProperties: false
      required: [protocolVersion, serverFeatures, sessionId, sessionToken, authKey, provisioningHostname, configRevision, syncIntervalSeconds, gatewayRoute, routes, wifiRoutes, config, expiresAt]
      properties:
        protocolVersion: {type: integer, const: 1}
        serverFeatures:
          type: array
          items: {type: string}
        sessionId: {type: string}
        sessionToken: {type: string, readOnly: true}
        authKey: {type: string, readOnly: true}
        provisioningHostname: {type: string}
        configRevision: {type: integer, minimum: 1}
        syncIntervalSeconds: {type: integer, minimum: 1}
        gatewayRoute: {type: string}
        routes:
          type: array
          items: {type: string}
        wifiRoutes:
          type: array
          items: {type: string}
        config: {$ref: "#/components/schemas/SessionConfig"}
        expiresAt: {$ref: "#/components/schemas/NullableTime"}
    Session:
      type: object
      required: [sessionId, gatewayRoute, routes, wifiRoutes, config, configRevision, appliedConfigRevision, deviceId, status, createdAt, expiresAt]
      properties:
        sessionId: {type: string}
        gatewayRoute: {type: string}
        routes:
          type: array
          items: {type: string}
        wifiRoutes:
          type: array
          items: {type: string}
        config: {$ref: "#/components/schemas/SessionConfig"}
        configRevision: {type: integer, minimum: 1}
        appliedConfigRevision: {type: integer, minimum: 0}
        deviceId: {type: string}
        status: {type: string, enum: [provisioning, active, cleaning, stopped, cleanup_failed]}
        createdAt: {type: string, format: date-time}
        expiresAt: {$ref: "#/components/schemas/NullableTime"}
    HistoricalSession:
      allOf:
        - $ref: "#/components/schemas/Session"
        - type: object
          required: [authKeyId, provisioningHostname, provisioningDeadline, lastSeenAt, syncDeadline, stoppedAt, stopReason, cleanupError]
          properties:
            authKeyId: {type: string}
            provisioningHostname: {type: string}
            provisioningDeadline: {$ref: "#/components/schemas/NullableTime"}
            lastSeenAt: {$ref: "#/components/schemas/NullableTime"}
            syncDeadline: {$ref: "#/components/schemas/NullableTime"}
            stoppedAt: {$ref: "#/components/schemas/NullableTime"}
            stopReason: {type: string}
            cleanupError: {type: string}
...